+++ packages/antd-ext/components/index.ts
@@ -1,8 +1,15 @@
 import ButtonExt from './ButtonExt';
 import ConfigProviderExt from './ConfigProviderExt';
+import TableExt from './TableExt';
 export * from 'antd';
+import zh_CN from 'antd/locale/zh_CN';
 
 export {
   ButtonExt,
   ConfigProviderExt,
+  TableExt,
+}
+
+export const locale = {
+  zh_CN,
 }