+++ packages/antd-ext/CHANGELOG.md
@@ -5,3 +5,8 @@
 ## 0.0.2-alpha.1
 1. 修正了ButtonExt hermes主题的样式
 2. 为组件层提供主题的单独配置
+
+## 0.0.3-alpha.1
+1. 新增 TableExt 组件，以及他的 hermes 主题
+2. 新增中文语言包的引用，导出为 locale
+3. 修正了 ButtonExt hermes 主题，link 样式不设置最小宽度